requirements.txt
pandas
PyPDF2
python-docx
python-pptx
tabula-py
pdfplumber
openpyxl
pillow
ebooklib
